multiagentsystems/requirements.txt
python-dotenv

crewai[tools]

langchain
langchain-community
langchain